+++ src/or/dirauth/shared_random_state.c
@@ -14,10 +14,11 @@
 #include "shared_random.h"
 #include "config.h"
 #include "confparse.h"
-#include "dirvote.h"
 #include "networkstatus.h"
 #include "router.h"
 #include "shared_random_state.h"
+
+#include "dirauth/dirvote.h"
 
 /* Default filename of the shared random state on disk. */
 static const char default_fname[] = "sr-state";
