+++ src/or/dirauth/shared_random_state.h
