+++ src/test/test_routerlist.c
@@ -19,7 +19,7 @@
 #include "container.h"
 #include "control.h"
 #include "directory.h"
-#include "dirvote.h"
+#include "dirauth/dirvote.h"
 #include "entrynodes.h"
 #include "hibernate.h"
 #include "microdesc.h"
@@ -30,7 +30,7 @@
 #include "routerlist.h"
 #include "routerset.h"
 #include "routerparse.h"
-#include "shared_random.h"
+#include "dirauth/shared_random.h"
 #include "statefile.h"
 #include "test.h"
 #include "test_dir_common.h"
