+++ src/test/test_shared_random.c
@@ -9,9 +9,9 @@
 #include "or.h"
 #include "test.h"
 #include "config.h"
-#include "dirvote.h"
-#include "shared_random.h"
-#include "shared_random_state.h"
+#include "dirauth/dirvote.h"
+#include "dirauth/shared_random.h"
+#include "dirauth/shared_random_state.h"
 #include "routerkeys.h"
 #include "routerlist.h"
 #include "router.h"
